+++ pkg/identityserver/store/store.go
@@ -291,7 +291,7 @@ var (
 
 	errAuthorizationNotFound     = errors.DefineNotFound("authorization_not_found", "authorization of `{user_id}` for `{client_id}` not found")
 	errAuthorizationCodeNotFound = errors.DefineNotFound("authorization_code_not_found", "authorization code not found")
-	errAccessTokenNotFound       = errors.DefineNotFound("access_token_not_found", "access token not found")
+	errAccessTokenNotFound       = errors.DefineNotFound("access_token_not_found", "access token `{access_token_id}` not found")
 
 	errAPIKeyNotFound = errors.DefineNotFound("api_key_not_found", "API key not found")